Zmena v parametrech tykajicich se skillu a par dalsich veci :) vice ci mene jsem musel sahnout na vsechny jiz hotovy skilly (i kdyz zas tak moc jich neni zejo :D) takze jejich autorum doporucuju aby se podivali co se zmenilo a pripadne i otestovali

+++ moriaengine/scripts/systems/skills/SkillUtils.cs
@@ -26,7 +26,7 @@ namespace SteamEngine.CompiledScripts {
 		public const int skillCheckVariance = 250;
 
 
-		public static bool CheckSuccess(ushort skillValue, int difficulty) {
+		public static bool CheckSuccess(int skillValue, int difficulty) {
 			// Chance to complete skill check given skill x and difficulty y
 			// ARGS:
 			//  skillValue = 0-1000
